chore: Refactor CI/CD workflows and enhance project scripts by adding Kubernetes setup, retagging for non-affected projects, and updating the pull request template. Modify backend API routes and frontend configurations for improved integration and deployment.
