--- apps/backend/config.py
@@ -1,3 +1,0 @@
-# Configuration for backend API
-
-API_PREFIX = "/api"
