+++ apps/frontend/next.config.ts
@@ -1,18 +1,7 @@
 import type { NextConfig } from "next";
 
-// Enable API proxy rewrites when backend host/port is available
-const backendHost = process.env.BACKEND_HOST || "localhost";
-const backendPort = process.env.BACKEND_PORT || "8000";
-
 const nextConfig: NextConfig = {
-  async rewrites() {
-    return [
-      {
-        source: "/api/:path*",
-        destination: `http://${backendHost}:${backendPort}/api/:path*`,
-      },
-    ];
-  },
+  // Rewrites are now handled by middleware.ts for runtime environment variable support
 };
 
 export default nextConfig;
